+++ src/ui/components/transactions/CreateOrder/CreateOrderCard.tsx
@@ -1,23 +1,23 @@
-import * as styles from './transfer.styl';
+import * as styles from './createOrder.styl';
 import * as React from 'react'
 import { translate, Trans } from 'react-i18next';
 import { TxIcon } from '../TransactionIcon';
 import { I18N_NAME_SPACE } from '../../../appConfig';
 import * as cn from 'classnames';
 import { OriginWarning } from '../OriginWarning';
-import { Balance } from '../../ui';
+import { Balance, Asset } from '../../ui';
 import { getMoney } from '../../../utils/converters';
-import { getAmount } from './parseTx';
+import { getAmount, getPrice } from './parseTx';
 
 @translate(I18N_NAME_SPACE)
-export class TransferCard extends React.PureComponent<ITransfer> {
+export class CreateOrderCard extends React.PureComponent<ICreateOrder> {
     
     render() {
         const className = cn(
-            styles.transferTransactionCard,
+            styles.createOrderTransactionCard,
             this.props.className,
             {
-                [styles.transferCard_collapsed]: this.props.collapsed
+                [styles.createOrderCard_collapsed]: this.props.collapsed
             },
         );
         
@@ -25,22 +25,37 @@ export class TransferCard extends React.PureComponent<ITransfer> {
         const { data = {} } = message;
         const tx = { type: data.type, ...data.data };
         const amount = getMoney(getAmount(tx), assets);
+        const price = getMoney(getPrice(tx), assets);
+        const isSell = tx.orderType === 'sell';
+        let iGet;
+        let sign = '- ';
+        if (!isSell) {
+            sign = '+ ';
+            iGet = amount;
+        } else {
+            iGet = amount.convertTo(price.asset, price.getTokens());
+        }
         
         return <div className={className}>
 
             <div className={styles.cardHeader}>
-                <div className={styles.transferTxIcon}>
+                <div className={styles.createOrderTxIcon}>
                     <TxIcon txType={this.props.txType}/>
                 </div>
                 <div>
                     <div className="basic500 body3 margin-min">
-                        <Trans i18nKey='transactions.transfer'>Send</Trans>
+                        {
+                            isSell ?
+                                <Trans i18nKey='transactions.orderSell'>Sell</Trans> :
+                                <Trans i18nKey='transactions.orderBuy'>Buy</Trans>
+                        }
+                        <span>: <Asset assetId={amount.asset.id}/>/<Asset assetId={price.asset.id}/></span>
                     </div>
                     <h1 className="headline1">
                         <Balance split={true}
-                                 addSign='- '
+                                 addSign={sign}
                                  showAsset={true}
-                                 balance={amount}
+                                 balance={iGet}
                                  className={styles.txBalanceWrapper} 
                         />
                     </h1>
@@ -57,7 +72,7 @@ export class TransferCard extends React.PureComponent<ITransfer> {
     }
 }
 
-interface ITransfer {
+interface ICreateOrder {
     assets: any;
     className: string;
     collapsed: boolean;