+++ src/ui/components/transactions/CreateOrder/CreateOrderFinal.tsx
@@ -1,25 +1,24 @@
-import './../pages/styles/transactions.styl';
 import * as React from 'react'
 import { translate, Trans } from 'react-i18next';
-import { I18N_NAME_SPACE } from '../../appConfig';
+import { I18N_NAME_SPACE } from '../../../appConfig';
 
 @translate(I18N_NAME_SPACE)
-export class BurnFinal extends React.PureComponent {
+export class CreateOrderFinal extends React.PureComponent {
     
     readonly props;
     
     render() {
-        const { tx, isApprove, isReject, isSend } = this.props;
-
+        const { isApprove, isReject, isSend } = this.props;
+        
         if (isApprove) {
             return <div>
                 <div className="margin-main headline2">
                     {isSend ? <Trans i18nKey='sign.transactionSend'>Your transaction is confirmed!</Trans> : null}
-                    {!isSend ? <Trans i18nKey='sign.transactionApproved'>Your transaction is signed</Trans> : null}
+                    {!isSend ? <Trans i18nKey='sign.transactionConfirmed'>Your transaction has been signed!</Trans>: null}
                 </div>
                 <div className="basic500">
-                    {isSend ? <span><Trans i18nKey='sign.transactionSendBurn'>You have burned</Trans> {tx.data.name}</span> : null}
-                    {!isSend ? <span><Trans i18nKey='sign.transactionConfirmedBurn'>You have approved a Burn transaction.</Trans></span> : null}
+                    {isSend ? <Trans i18nKey='sign.transactionSendAlias'>An alias has been created.</Trans> : null}
+                    {!isSend ? <Trans i18nKey='sign.transactionConfirmAlias'>You have approved Alias creation.</Trans> : null}
                 </div>
             </div>
         }